Begin implementing automatic config yanking
