+++ src/market.rs
@@ -30,28 +30,6 @@ pub async fn log_blocks(blocks: Vec<serde_json::Value>) -> Result<(), String> {
     Ok(())
 }
 
-#[derive(Clone, PartialEq, Debug, serde::Serialize, serde::Deserialize)]
-pub struct Sale {
-    pub price: u64,
-    pub market_name: String,
-}
-impl Sale {
-    pub fn from_item(i: &hacksteader::Item) -> Result<Self, hacksteader::AttributeParseError> {
-        use hacksteader::AttributeParseError::*;
-
-        Ok(Sale {
-            market_name: i
-                .get("market_name")
-                .ok_or(MissingField("market_name"))?
-                .s
-                .as_ref()
-                .ok_or(WronglyTypedField("market_name"))?
-                .clone(),
-            price: i.get("price")?.n.as_ref()?.parse().ok()?,
-        })
-    }
-}
-
 pub async fn market_search(
     db: &DynamoDbClient,
     cat: Category,
